refactor: spec 내부 엔티티들 같은 계층으로 분리
