+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/activitynetworking/entity/ActivityNetworking.java
@@ -1,16 +1,14 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.entity;
+package kakaotech.bootcamp.respec.specranking.domain.activitynetworking.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kakaotech.bootcamp.respec.specranking.domain.common.type.CareerRole;
+import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Spec;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,7 +16,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class WorkExperience {
+public class ActivityNetworking {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,20 +27,19 @@ public class WorkExperience {
     @JoinColumn(name = "spec_id", nullable = false, columnDefinition = "BIGINT UNSIGNED")
     private Spec spec;
 
-    @Column(name = "company_name", nullable = false, columnDefinition = "VARCHAR(50)")
-    private String companyName;
+    @Column(name = "activity_name", nullable = false, columnDefinition = "VARCHAR(50)")
+    private String activityName;
 
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, columnDefinition = "VARCHAR(20) DEFAULT 'INTERN'")
-    private CareerRole position;
+    @Column(nullable = false, columnDefinition = "VARCHAR(50)")
+    private String position;
 
-    @Column(name = "work_month", nullable = false, columnDefinition = "SMALLINT UNSIGNED")
-    private Integer workMonth;
+    @Column(columnDefinition = "VARCHAR(50)")
+    private String award;
 
-    public WorkExperience(Spec spec, String companyName, CareerRole position, Integer workMonth) {
+    public ActivityNetworking(Spec spec, String activityName, String position, String award) {
         this.spec = spec;
-        this.companyName = companyName;
+        this.activityName = activityName;
         this.position = position;
-        this.workMonth = workMonth;
+        this.award = award;
     }
 }
