+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/activitynetworking/repository/ActivityNetworkingRepository.java
@@ -1,6 +1,6 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
+package kakaotech.bootcamp.respec.specranking.domain.activitynetworking.repository;
 
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.ActivityNetworking;
+import kakaotech.bootcamp.respec.specranking.domain.activitynetworking.entity.ActivityNetworking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ActivityNetworkingRepository extends JpaRepository<ActivityNetworking, Long> {
