+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/certification/repository/CertificationRepository.java
@@ -1,6 +1,6 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
+package kakaotech.bootcamp.respec.specranking.domain.certification.repository;
 
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Certification;
+import kakaotech.bootcamp.respec.specranking.domain.certification.entity.Certification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CertificationRepository extends JpaRepository<Certification, Long> {
