+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/education/entity/Education.java
@@ -1,4 +1,4 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.entity;
+package kakaotech.bootcamp.respec.specranking.domain.education.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,6 +12,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.FinalEducation;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.FinalStatus;
+import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Spec;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
