+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/education/repository/EducationRepository.java
@@ -1,6 +1,6 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
+package kakaotech.bootcamp.respec.specranking.domain.education.repository;
 
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Education;
+import kakaotech.bootcamp.respec.specranking.domain.education.entity.Education;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EducationRepository extends JpaRepository<Education, Long> {
