+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/educationdetail/repository/EducationDetailRepository.java
@@ -1,6 +1,6 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
+package kakaotech.bootcamp.respec.specranking.domain.educationdetail.repository;
 
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.EducationDetail;
+import kakaotech.bootcamp.respec.specranking.domain.educationdetail.entity.EducationDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EducationDetailRepository extends JpaRepository<EducationDetail, Long> {
