+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/englishskill/entity/EnglishSkill.java
@@ -1,6 +1,14 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.entity;
+package kakaotech.bootcamp.respec.specranking.domain.englishskill.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Spec;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
