+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/englishskill/repository/EnglishSkillRepository.java
@@ -1,6 +1,6 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
+package kakaotech.bootcamp.respec.specranking.domain.englishskill.repository;
 
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.EnglishSkill;
+import kakaotech.bootcamp.respec.specranking.domain.englishskill.entity.EnglishSkill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EnglishSkillRepository extends JpaRepository<EnglishSkill, Long> {
