--- src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/repository/PortfolioRepository.java
@@ -1,7 +1,0 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
-
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Portfolio;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
-}
