+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/service/SpecService.java
@@ -1,33 +1,33 @@
 package kakaotech.bootcamp.respec.specranking.domain.spec.service;
 
 import java.util.Optional;
+import kakaotech.bootcamp.respec.specranking.domain.activitynetworking.entity.ActivityNetworking;
+import kakaotech.bootcamp.respec.specranking.domain.activitynetworking.repository.ActivityNetworkingRepository;
 import kakaotech.bootcamp.respec.specranking.domain.ai.dto.mapping.AiDtoMapping;
 import kakaotech.bootcamp.respec.specranking.domain.ai.dto.request.AiPostSpecRequest;
 import kakaotech.bootcamp.respec.specranking.domain.ai.dto.response.AiPostSpecResponse;
 import kakaotech.bootcamp.respec.specranking.domain.ai.service.AiService;
+import kakaotech.bootcamp.respec.specranking.domain.certification.entity.Certification;
+import kakaotech.bootcamp.respec.specranking.domain.certification.repository.CertificationRepository;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.CareerRole;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.Degree;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.FinalEducation;
 import kakaotech.bootcamp.respec.specranking.domain.common.type.FinalStatus;
+import kakaotech.bootcamp.respec.specranking.domain.education.entity.Education;
+import kakaotech.bootcamp.respec.specranking.domain.education.repository.EducationRepository;
+import kakaotech.bootcamp.respec.specranking.domain.educationdetail.entity.EducationDetail;
+import kakaotech.bootcamp.respec.specranking.domain.educationdetail.repository.EducationDetailRepository;
+import kakaotech.bootcamp.respec.specranking.domain.englishskill.entity.EnglishSkill;
+import kakaotech.bootcamp.respec.specranking.domain.englishskill.repository.EnglishSkillRepository;
 import kakaotech.bootcamp.respec.specranking.domain.spec.dto.request.PostSpecRequest;
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.ActivityNetworking;
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Certification;
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Education;
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.EducationDetail;
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.EnglishSkill;
 import kakaotech.bootcamp.respec.specranking.domain.spec.entity.Spec;
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.WorkExperience;
-import kakaotech.bootcamp.respec.specranking.domain.spec.repository.ActivityNetworkingRepository;
-import kakaotech.bootcamp.respec.specranking.domain.spec.repository.CertificationRepository;
-import kakaotech.bootcamp.respec.specranking.domain.spec.repository.EducationDetailRepository;
-import kakaotech.bootcamp.respec.specranking.domain.spec.repository.EducationRepository;
-import kakaotech.bootcamp.respec.specranking.domain.spec.repository.EnglishSkillRepository;
 import kakaotech.bootcamp.respec.specranking.domain.spec.repository.SpecRepository;
-import kakaotech.bootcamp.respec.specranking.domain.spec.repository.WorkExperienceRepository;
 import kakaotech.bootcamp.respec.specranking.domain.store.service.FileStore;
 import kakaotech.bootcamp.respec.specranking.domain.user.entity.User;
 import kakaotech.bootcamp.respec.specranking.domain.user.repository.UserRepository;
 import kakaotech.bootcamp.respec.specranking.domain.user.util.UserUtils;
+import kakaotech.bootcamp.respec.specranking.domain.workexperience.entity.WorkExperience;
+import kakaotech.bootcamp.respec.specranking.domain.workexperience.repository.WorkExperienceRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
