+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/workexperience/repository/WorkExperienceRepository.java
@@ -1,6 +1,6 @@
-package kakaotech.bootcamp.respec.specranking.domain.spec.repository;
+package kakaotech.bootcamp.respec.specranking.domain.workexperience.repository;
 
-import kakaotech.bootcamp.respec.specranking.domain.spec.entity.WorkExperience;
+import kakaotech.bootcamp.respec.specranking.domain.workexperience.entity.WorkExperience;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WorkExperienceRepository extends JpaRepository<WorkExperience, Long> {
